+++ Packages/com.shelly.scenetransition/Runtime/Scripts/WorkflowContext.cs
@@ -1,0 +1,14 @@
+namespace Shelly.SceneTransition
+{
+    public class WorkflowContext
+    {
+        public readonly SceneTransitionBehaviour FaderPrefab;
+
+        public SceneTransitionBehaviour CurrentFader;
+
+        public WorkflowContext(SceneTransitionBehaviour faderPrefab)
+        {
+            FaderPrefab = faderPrefab;
+        }
+    }
+}